+++ activity/auth/AuthActivity.js
@@ -1,46 +1,17 @@
 import React from 'react';
-import { Asset } from 'expo-asset';
+import AppLoading from 'expo-app-loading';
 import { StyleSheet,LogBox,useColorScheme } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import * as Font from 'expo-font';
 import 'moment/locale/id';
-import {
-	Inter_300Light,
-	Inter_400Regular,
-	Inter_500Medium,
-	Inter_600SemiBold,
-	Inter_700Bold,
-} from '@expo-google-fonts/inter';
 import AppNavigator from './AppNavigator';
 import { AuthProvider } from './AuthProvider';
-import {AppearanceProvider} from 'react-native-appearance'
-import remoteConfig from '@react-native-firebase/remote-config'
+import loadResources from '@pn/utils/Assets'
 
 LogBox.ignoreLogs(['Possible Unhandled Promise Rejection']);
 
 async function loadResourcesAsync() {
 	// load all resources such as images, fonts, etc.
-	await Promise.all([
-		Asset.loadAsync([
-			require('../../assets/icon.png'),
-			require('../../assets/splash.png'),
-			require('../../assets/avatar.png'),
-			require('../../assets/landing.png'),
-			require('../../assets/transparent.png'),
-			require('../../assets/login.png'),
-			require('../../assets/forget.png'),
-			require('../../assets/register.png'),
-		]),
-		Font.loadAsync({
-			Inter_300Light,
-			Inter_Regular: Inter_400Regular,
-			Inter_Medium:Inter_500Medium,
-			Inter_SemiBold: Inter_600SemiBold,
-			Inter_Bold: Inter_700Bold,
-		}),
-		remoteConfig().setDefaults({tuner_enabled:false})
-		.then(()=>remoteConfig().fetchAndActivate())
-	]);
+	await loadResources();
 }
 
 const styles = StyleSheet.create({
@@ -50,18 +21,35 @@ const styles = StyleSheet.create({
 });
 
 export default function Auth() {
-    React.useEffect(()=>{
-        loadResourcesAsync();
-    },[])
+	const [isLoadingComplete, setLoadingComplete] = React.useState(false);
+
+	if (!isLoadingComplete) {
+		return (
+			<AppLoading
+				startAsync={loadResourcesAsync}
+				onError={(err)=>handleLoadingError(err,setLoadingComplete)}
+				onFinish={() => handleFinishLoading(setLoadingComplete)}
+			/>
+		);
+	}
 
 	return (
         <SafeAreaView style={[styles.container]}>
-            <AppearanceProvider>
-                <AuthProvider>
-                    <AppNavigator />
-                </AuthProvider>
-            </AppearanceProvider>
-            
+			<AuthProvider>
+				<AppNavigator />
+			</AuthProvider>
         </SafeAreaView>
     );
+}
+
+function handleLoadingError(error,setLoadingComplete) {
+	// In this case, you might want to report the error to your error reporting
+	// service, for example Sentry
+	console.log(error);
+	setLoadingComplete(true);
+}
+
+function handleFinishLoading(setLoadingComplete) {
+	setLoadingComplete(true);
+
 }
