+++ components/navigation/Header.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {Divider} from '@ui-kitten/components'
 import {Animated,NativeScrollEvent,NativeSyntheticEvent} from 'react-native'
 import {useNavigationState} from '@react-navigation/native'
-import {AuthContext} from '@pn/provider/AuthProvider'
+import {AuthContext} from '@pn/provider/Context'
 import LottieView,{AnimatedLottieViewProps} from 'lottie-react-native'
 import TopNavigationAction from './TopAction'
 import TopNav,{TopNavigationProps} from './TopNav';
