+++ index.js
@@ -1,7 +1,7 @@
 import 'react-native-gesture-handler'
 import 'expo-asset'
 import {AppRegistry} from 'react-native'
-import App from './App';
+import App from './provider/AuthProvider';
 import Auth from './activity/auth/AuthActivity'
 import Share from './activity/share/ShareActivity'
 import SyncAdapter from './services/SyncService'
