+++ screens/Geodata/Transform.js
@@ -15,7 +15,7 @@ import style from '@pn/components/global/style'
 import Button from '@pn/components/global/Button'
 import Pagination from '@pn/components/global/Pagination'
 import useClipboard from '@pn/utils/clipboard'
-import { AuthContext } from '@pn/provider/AuthProvider';
+import { AuthContext } from '@pn/provider/Context';
 import {randomInt} from '@pn/utils/Main'
 import ListItem from '@pn/components/global/ListItem'
 import verifyRecaptcha from '@pn/module/Recaptcha'
