+++ screens/Tools/ImagesChecker.js
@@ -13,7 +13,7 @@ import {AdsBanner,AdsBanners,showInterstisial} from '@pn/components/global/Ads'
 import useAPI from '@pn/utils/API'
 import style from '@pn/components/global/style'
 import Button from '@pn/components/global/Button'
-import { AuthContext } from '@pn/provider/AuthProvider';
+import { AuthContext } from '@pn/provider/Context';
 import { ucwords,extractMeta,randomInt } from '@pn/utils/Main';
 import verifyRecaptcha from '@pn/module/Recaptcha'
 import { convertFile, pickImage } from '@pn/utils/PickLibrary';
